refactor: refactor hooks to use named export
