+++ jest.config.js
@@ -15,15 +15,25 @@ const customJestConfig = {
   modulePathIgnorePatterns: ["<rootDir>/integration-test/", "<rootDir>/.next/"],
   testPathIgnorePatterns: ["<rootDir>/node_modules/"],
   moduleNameMapper: {
+    "^@/components": "<rootDir>/src/components",
     "^@/components/(.*)$": "<rootDir>/src/components/$1",
+    "^@/utils": "<rootDir>/src/utils",
     "^@/utils/(.*)$": "<rootDir>/src/utils/$1",
+    "^@/types": "<rootDir>/src/types",
     "^@/types/(.*)$": "<rootDir>/src/types/$1",
+    "^@/hooks": "<rootDir>/src/hooks",
     "^@/hooks/(.*)$": "<rootDir>/src/hooks/$1",
+    "^@/style": "<rootDir>/src/style",
     "^@/style/(.*)$": "<rootDir>/src/style/$1",
+    "^@/services": "<rootDir>/src/services",
     "^@/services/(.*)$": "<rootDir>/src/services/$1",
+    "^@/lib": "<rootDir>/src/lib",
     "^@/lib/(.*)$": "<rootDir>/src/lib/$1",
+    "^@/pages": "<rootDir>/src/pages",
     "^@/pages/(.*)$": "<rootDir>/src/pages/$1",
+    "^@/mocks": "<rootDir>/src/mocks",
     "^@/mocks/(.*)$": "<rootDir>/src/mocks/$1",
+    "^@/contexts": "<rootDir>/src/contexts",
     "^@/contexts/(.*)$": "<rootDir>/src/contexts/$1",
     "react-markdown": "<rootDir>/src/lib/test/react-markdown.jsx",
     "remark-frontmatter": "<rootDir>/src/lib/test/remark-frontmatter.js",
