+++ src/components/source/ConfigureSourceForm.tsx
@@ -17,7 +17,7 @@ import { useDeleteSource } from "@/services/connector";
 import { useAmplitudeCtx } from "@/contexts/AmplitudeContext";
 import { sendAmplitudeData } from "@/lib/amplitude";
 import { ErrorDetails, Violation } from "@/lib/instill/types";
-import useDeleteResourceGuard from "@/hooks/useDeleteResourceGuard";
+import { useDeleteResourceGuard } from "@/hooks/useDeleteResourceGuard";
 
 export type ConfigureSourceFormProps = {
   source: Nullable<SourceWithPipelines>;
