+++ src/pages/pipelines/[id].tsx
@@ -20,7 +20,7 @@ import {
 } from "@/components/ui";
 import { ConfigurePipelineForm } from "@/components/pipeline";
 import { useAmplitudeCtx } from "@/contexts/AmplitudeContext";
-import { useSendAmplitudeData } from "@/hooks/useSendAmplitudeData";
+import { useSendAmplitudeData } from "@/hooks";
 import { Pipeline } from "@/lib/instill";
 import { getCodeHikeTemplateSource } from "@/lib/markdown";
 import { MDXRemote, MDXRemoteSerializeResult } from "next-mdx-remote";
