+++ tsconfig.json
@@ -2,15 +2,25 @@
   "compilerOptions": {
     "baseUrl": "./src",
     "paths": {
+      "@/components": ["components"],
       "@/components/*": ["components/*"],
+      "@/types": ["types"],
       "@/types/*": ["types/*"],
+      "@/hooks": ["hooks"],
       "@/hooks/*": ["hooks/*"],
+      "@/utils": ["utils"],
       "@/utils/*": ["utils/*"],
+      "@/style": ["styles"],
       "@/style/*": ["styles/*"],
+      "@/services": ["services"],
       "@/services/*": ["services/*"],
+      "@/lib": ["lib"],
       "@/lib/*": ["lib/*"],
+      "@/pages": ["pages"],
       "@/pages/*": ["pages/*"],
+      "@/mocks": ["mocks"],
       "@/mocks/*": ["mocks/*"],
+      "@/contexts": ["contexts"],
       "@/contexts/*": ["contexts/*"]
     },
     "target": "es5",
